Convert all parameters that use enums to use that stronger type
Many system parameters use values defined in an enum but are manipulated as ints which is not type safe. This changes each enum to C++11 enum class for greatest type safety and converts each use of the define to use the correct type. No bugs were found.

Enums used as bitfields are left unchanged for now.

Tests:
 - Build and boot