Split the PwmGenerator monolithic class up
As the first step of allowing openinverter to be more
portable and testable split the monolithic PwmGenerator
class and FOC/Sine implementations up into separate
classes.

The architecture splits the main motor control algorithm
into 3 template classes:
 - PwmGeneratorBase - for all common code
 - SinePwmGenerator - for the original Sine motor control
 - FocPwmGenerator - for Field Oriented Control algorithm
The classes are implement static polymorphism using the
Curiously Recurring Template Pattern (CRTP) and all code is
now contained in the respective headers. It may be possible
to split the code off into separate compliation units but as
further refactoring is likely to be required keeping code and
interface together should make changes easier to follow.

The STM32 platform specific code for PWM is now
contained within the STM32PwmDriver class.

In future it will be possible to reuse the SinePwmGenerator
and FocPwmGenerator in test code or other platforms but
futher dependencies exist (AnaIn, Encoder and possibly
SineCore and Params).

Tests:
 - Compile both Sine and FOC builds