Modularise build system and isolate platform specific
This change introduces a PLATFORM cache property that
can be used externally to cmake to specify the build
platform. At the moment only the "host" platform is
present. If not specified the "host" platform is automatically
selected.

All platform specific build code is now isolated in the
"common-build-options" tree and in the global_options and
platform_options CMake interface libraries it contains.

The unit tests are a special case and are conditionally
included at the top-level when the "host" platform is
specified.

Tests:
 - Build and verify the unit test builds, can be debugged and
   run
 - Verify the coverage_OpenInverterTest target works and
   generates updated code coverage