Create STM32F1 platform in CMake build
This adds a CMake platform for the STM32F103 equivalent
to the existing stm32_sine Makefile build. No changes are
made to source code location. The STM32 platform specific
code is included in a new libopeninv_stm32f1 library and
platform exectuable.

The CMake build process has a custom toolchain definition
for the STM32F1 processor and this assumes that the
arm-none-eabi-* toolchain is on the path.

Changed CMake files formatted with cmake-format.

Tests:
 - Configure and build with:
 mkdir build/stm32f1
 cd build/stm32f1
 cmake -GNinja -DCMAKE_BUILD_TYPE=Release -DPLATFORM=stm32f1 ../..
 cmake --build .
 - Run resulting build with the new VS Code debug target
   "Debug CMake STM32F1"