Add CMakePresets for Linux and STM32F1
A CMakePresets.json integrates nicely with Visual Studio
Code and allows GUI selection of the normal Linux/host
platform or the STM32F1 platform. A build preset is used for
building and gathering code coverage of the unit tests. The
default build preset works fine for the other targets.

Tests:
 - Configure and build each configure target
 - Run the Unit Test Coverage build preset on the host
   platform
 - Run a reconfigure and rebuild from the VS Code CMake UI
 - Delete the build directory and verify all targets can be
   rebuilt