Add C2000 build target
This adds the C2000 as a build target. The CMake
configuration process assumes that the compiler is available
on the path (as the STM32F1). A basic set of compiler and
linker options are provided that match the TI Code
Composer Studio prototype project.

The build fails to compile but does not affect other builds.