Add a skelton inverter
This create a skeleton inverter application which includes
an outline implementation of the Field Oriented Control
PWM generation class. The required analog in, shaft
encoder and PWM driver are all empty of functionality at
this point.

The .bss section is too small so extended it some more with
another global RAM block.

Tests:
 - Build all platforms
 - Run inverter on the TI launchpad and verify it does nothing
   untoward other than get stuck in an infinite loop.