Add pwmsimulator cross-platform test application
pwmsimulator is a stub application to be filled out. It builds
for both C2000 and the host platform but not STM32F1 due
to a missing C library.

The C2000 custom entry point link command is moved to
the device_support library where the entry point code lives.
It is not required for a basic application.

A new link_last library is defined for the host platform and
C2000. This should be added last to any executable link
libraries to ensure any necessary libraries are linked last.
Funny that! ;-)

Tests:
 - Build and run and debug pwmsimulator on x86 Linux and
   C2000 launchpad.