Fix integer overflow in FOC constant and sqrt calculation
The modMax constant of 2/sqrt(3) in the FOC calculation
has a signed integer overflow. This results in undefined
behaviour and ends up with a different on C2000/x86_64
compared to ARM. The fix is to evaluate in an external
calulator and store as a floating point constant similar to the
other constants for FOC.

The fpsqrt method was forcing all input values to be
INT32_MAX rather than clamping values above this to
INT32_MAX. Fixed by using MIN() rather than MAX().

Tests:
 - Run against new FOC test suite