New test suite for FOC
Add a new test suite to excercise the core FOC calculations. Most tests cross-check against flloating point implementations of the calculations allowing the maths to be checked.

Host C++ standard uped to C++20 to allow the handy
numeric constants in <utility> to be used.

CMakeLists.txt in unit test is tidied up and all sources are
now explicitly listed as per CMake best-practice.

Pull in libopeninv with fixes identified by the tests

Tests:
 - Run new unit tests.
 - TestFocPwmGeneration.NormalModeRun and
    .ManualModeRun tests now fail. Fixes will follow.