Fix TestFocPwmGeneration.NormalModeRun and .ManualModeRun tests
These tests were broken by the fixes to FOC in df9638. The
failures are due to poor tests.

ManualModeRun is fixed by increasing the manualiq
supplied to ensure a greater PWM modulation. Also, the
encoder angle is changed to 10,000 (approx 54.9 degrees)
to ensure that Id is not aligned with any of the stator phase
angles and thus not require any PWM modulation.

NormalModeRun is fixed in a simular way but the throttle value is fixed to be 50%. This is achieved by setting to the value 50 rather than 0.5 as previously assumed.

Tests:
 - Run all tests and verify that they pass