Fix FOC FP constants to avoid overflow
The foc module increases CST_DIGITS to 15. When using
FP_FROMINT with constants this will cause the C2000 to
experience an integer overflow as constants are int unless
specified. Explicitly specify constants as long to avoid the
overflow. This should cause any issues on 32-bit ARM and
x86_64 systems.

Tests:
 - Build on all platforms. Verify unit tests still pass on Linux.
 - No tests on C2000 yet.