Enable asserts on debug builds of C2000
This change defines the DEBUG define on the C2000
platform. This will cause ASSERT statements in driverlib to
be evaluated. To do this the __error__() function signature is
updated to take const char * for the filename parameter to
avoid problems when building C++ modules that use the
driverlib headers.

The C2000 CPU1 and _LAUNCHXL_F28379 defines have
been moved to the common platform_options CMake target
as they are required everywhere that the device.h header is
used not just the device_support library.

Tests:
 - Build and run inverter on the C2000. Verify nothing bad
   happens.