Document current state of the firmware
This change includes details on the current state and
direction of the firmware in a new README.md.

The CMake preset for C2000 is now c2000 to match the
case of the other presets.

Details of the JTAG interface necessary for programming
the Tesla M3 inverter are also included.

Tests:
 - Check spelling
 - Review in Firefox with a markdown parser