Tidy documentation on setting up a CCS debug project
Fix some problems reported by Damien Maguire when trying to set up a Code Composer Studio project. Turns out it is better to create the CMake build project outside of the CCS workspace directory and import it.

Clarify which parts of code the TI license correspond to.

Fix the spacing of Model 3 / Y.

Tests:
 - None