Auto detect operation on Tesla M3 inverter vs Launchpad
Use the pin count of the C2000 to detect whether we are
running on the Tesla M3 inverter or a LAUNCHXL-F28379D
eval board. The Tesla board uses a 176 pin QFP part but the
eval board uses a 337 pin BGA.

Change the blinky test application to flip which GPIO pins
are used for the LEDs.

Remove the _LAUNCHXL_F28379D define as we are now
detecting HW dynamically at runtime.

Tests:
 - Run on Tesla M3 inverter and LAUNCHXL-F28379D and
    verify the LEDs blink.