Fix operation of gate driver test
This changeset refactors the HW specific
parts of the STGAP1AS gate driver
interface into a separate class to avoid
cluttering the logic of configuring the
STGAP1AS.

Fix a bug in
TeslaM3GateDriver::WriteRegister()
which sent the wrong data causing the
write to be ignored by the chips.

Change the method of verifying the gate
driver configuration to account for the
behaviour of the STGAP1AS to include
"don't care" bits in configuration
registers in the CRC calculation and for
these bits varying from read to read.
Each register configuration now requires
a mask to be specified for which bits in a
register contain valid data. These bits
come from the STGAP1AS datasheet.

Temporary debug printf code remains.

A typo in the stgap1as_gate_driver.h
register definitions has been fixed.

Tests:
 - Run against M3 inverter with gate PSU
   enabled and disabled.
 - Verify that with the gate PSU enable
   that configuration is successful and
   status reports OK when polled.
 - With the gate drive PSU disabled verify
   that the errors reported correspond to
   remote registers in the STGAP1AS
   chips which cannot be set without the
   isolated supply on the chip being
   powered up.
 - Verify operation with a logic analyser
   connected to the SPI bus. Cross-check
   with traces taken from an inverter
   running Tesla firmware.