Integate gate driver support into main C2000 inverter
Update the c2000 inverter application to use the gate driver.
This involves turning on the gate drive PSU before initialising
the gate drivers for each phase. The fault state of the gate
drivers is reported to stdout every second.

Add a simple heartbeat LED toggle to the main loop to
compensate for the gate drive fault LED no longer being on (!)
as a way of determinig if the inverter is powered up.

Tests:
 - Run on the Tesla M3 inverter and verify it operates with the
   gate drive fault LED extinguished and not gate drive fault is
   reported.