Fix PMIC state transition failure on initialisation
When the PMIC state transition is sampled at the end of the
configuration process it reports STBYEN in the DEVSTAT
register indicating that the standby LDO is still enabled. This
doesn't really matter so just mask off and check the overall
PMIC state and ignore the LDO enable states.

Update the PMIC test program to report the success or failure
of initialisation.

Tests:
 - Run updated test program before fixing the state check

Reported-by: Bernd Ocklin <bernd@ocklin.de>