Allow PWM when debugging
This allows the phase and resolver PWM to continue
operating when debugging. In particular this stops a nasty
interaction between the PWM control loop and printf() in the
main loop. The PWM interrupts are blocked while lines from
stdout are relayed back to the host debugger causing weird
glitches in the PWM and resolver exciter output.

Tests:
 - Run on Tesla M3 inverter and monitor PWM waveform and
   resolver exciter. Observe no glitching and jumping in time
   to the printf output every 1/2 second.