Capture key motor control analog signals
This change configures the 4 channels of C2000 ADC to fire
simultaneously at the centre of the phase PWM waveform.
The channels are configured to pick up the Phase A and Phase
B currents and resolver sine and cosine signals.

The ADC integration is contained within the
MotorAnalogCapture class. The PwmDriver is updated to
start ADC conversions and the main control loop is refactored
to run on completion of the ADC conversion rather than an
event directly from the PWM. It seems difficult to eliminate
coupling between the ADC capture and PWM driver but these
classes try to isolate much of the details and only the high-
level triggering is shared.

The inverter application is updated to print out the values
captured. The readings are not integrated into any other part
of the system at this point.

Tests:
 - Run on Tesla Model 3 inverter. Verify that the PWM is still
   generated and the main control loop runs as before.
 - Check that phase current values are approx mid-scale
 - With a resolver attached observe that moving an iron core
   near the sine and cosine windings vary the correspoding
   readings
 - Run on the Launchxl board and verify zero readings when
   nothing connected to each analog input