Fix ADC triggering when Phase A fully on or off
In commit 27773a when Phase A is fully on or off it no longer
triggers the ADC capture stalling the entire motor control
loop. This fixes that by using EPWM4 which was previously
only used as a synchronisation pass-through for that
purpose. This explains the EPWM topology that Tesla
appears to use.

The Launchxl EPWM is rejigged to use the same EPWM4 for
ADC capture as the Tesla inverter. We can't use the same
EPWM topology as the Tesla inverter because EPWM7 is not
broken out in an accessible way.

Tests:
 - Run on Tesla M3 inverter and Launchxl and verify the PWM
   waveforms and ADC sampling continue to operate
 - Force Phase A to be set to 4096 and 0 and verify ADC
   sampling and main motor control loop continues to operate