Add safety warning to the README
After a recent confusion where an inexperienced user bought a JTAG adapter and was trying to flash their inverter expecting the software to be functional put a warning up front. It may save someone from accidentally breaking their inverter or being misled into thinking the project is further on than it is.